third_party/WebKit/LayoutTests/inspector/components/file-path-scoring-expected.txt

Test file path scoring function

Expected score must be equal to the actual score
Test set size: 588

=== Test query: textepl
    expected return: ./Source/devtools/front_end/>TextE<ditor.>pl< (score: 176)
      actual return: ./Source/devtools/front_end/>TextE<ditor.>pl< (score: 176)

=== Test query: defted
    expected return: ./Source/devtools/front_end/>Def<ault>T<ext>Ed<itor.pl (score: 134)
      actual return: ./Source/devtools/front_end/>Def<ault>T<ext>Ed<itor.pl (score: 134)

=== Test query: CMTE
    expected return: ./Source/devtools/front_end/>C<ode>M<irror>T<ext>E<ditor.pl (score: 109)
      actual return: ./Source/devtools/front_end/>C<ode>M<irror>T<ext>E<ditor.pl (score: 109)

=== Test query: frocmte
    expected return: ./Source/devtools/>fro<nt_end/>C<ode>M<irror>T<ext>E<ditor.pl (score: 138)
      actual return: ./Source/devtools/>fro<nt_end/>C<ode>M<irror>T<ext>E<ditor.pl (score: 138)

=== Test query: cmtepl
    expected return: ./Source/devtools/front_end/>C<ode>M<irror>T<ext>E<ditor.>pl< (score: 117)
      actual return: ./Source/devtools/front_end/>C<ode>M<irror>T<ext>E<ditor.>pl< (score: 117)

=== Test query: setscr
    expected return: ./Source/devtools/front_end/>Set<tings>Scr<een.pl (score: 137)
      actual return: ./Source/devtools/front_end/>Set<tings>Scr<een.pl (score: 137)

=== Test query: cssnfv
    expected return: ./Source/devtools/front_end/>CSSN<amed>F<low>V<iew.pl (score: 147)
      actual return: ./Source/devtools/front_end/>CSSN<amed>F<lowCollections>V<iew.pl (score: 147)

=== Test query: jssf
    expected return: ./Source/devtools/front_end/>J<ava>S<cript>S<ource>F<rame.pl (score: 85)
      actual return: ./Source/devtools/front_end/>J<ava>S<cript>S<ource>F<rame.pl (score: 85)

=== Test query: sofrapl
    expected return: ./Source/devtools/front_end/>So<urce>Fra<me.>pl< (score: 142)
      actual return: ./Source/devtools/front_end/>So<urce>Fra<me.>pl< (score: 142)

=== Test query: inspeins
    expected return: ./Source/core/inspector/>Inspe<ctor>Ins<trumentation.bpp (score: 203)
      actual return: ./Source/core/inspector/>Inspe<ctorCanvas>Ins<trumentation.z (score: 203)

=== Test query: froscr
    expected return: ./Source/devtools/>fro<nt_end/>Scr<ipt.pl (score: 131)
      actual return: ./Source/devtools/>fro<nt_end/>Scr<ipt.pl (score: 131)

=== Test query: adscon
    expected return: ./Source/devtools/front_end/>Ad<vanced>S<earch>Con<troller.pl (score: 129)
      actual return: ./Source/devtools/front_end/>Ad<vanced>S<earch>Con<troller.pl (score: 129)

=== Test query: execontext
    expected return: dom/>Exec<utionC>ontext<.cpp (score: 253)
      actual return: dom/>Exec<utionC>ontext<.cpp (score: 253)
